% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdclient.r
\name{pbdclient}
\alias{pbdclient}
\title{Client Launcher}
\usage{
pbdclient(addr = "localhost", port = 55555, prompt = "remoter")
}
\arguments{
\item{addr, port, prompt}{See \code{?remoter::client}}
}
\description{
Connect to a pbdR server (launch the client).
}
\details{
An alias of the remoter client connecter function.
}

