% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdserver.r
\name{pbdserver}
\alias{pbdserver}
\title{Server Launcher}
\usage{
pbdserver(port = 55555, remote_port = 55556, bcaster = "zmq",
  password = NULL, maxretry = 5, secure = has.sodium(), log = TRUE,
  verbose = FALSE, showmsg = FALSE)
}
\arguments{
\item{port}{The port (number) that will be used for communication between 
the client and server.  The port value for the client and server
must agree.}

\item{remote_port}{TODO}

\item{bcaster}{TODO}

\item{password}{A password the client must enter before the user can process
commands on the server.  If the value is \code{NULL}, then no
password checking takes place.}

\item{maxretry}{The maximum number of retries for passwords before shutting
everything down.}

\item{secure}{Logical; enables encryption via public key cryptography of
the 'sodium' package is available.}

\item{log}{Logical; enables some basic logging in the server.}

\item{verbose}{Logical; enables the verbose logger.}

\item{showmsg}{Logical; if TRUE, messages from the client are logged}
}
\value{
Returns \code{TRUE} invisibly on successful exit.
}
\description{
Launcher for the pbdCS server.  The launcher should be executed
in batch using something like \code{mpirun}.
}
\details{
TODO
}

