% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdclient.r
\name{pbdclient}
\alias{pbdclient}
\title{Client Launcher}
\usage{
pbdclient(prompt = "pbdR", ...)
}
\arguments{
\item{prompt}{The prompt to display differentiating the pbd client
from a basic R prompt.}

\item{...}{Additional arguments passed to \code{remoter::client}.
See \code{?remoter::client} for details.}
}
\description{
Connect to a pbdR server (launch the client).
}
\details{
An alias of the remoter function \code{client()}.
}

